Refactor: export all sidebar together in index.tsx
