+++ src/components/Layout/IdeaLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { IdeaBar } from "../Sidebar/IdeaBar";
+import { IdeaBar } from "../Sidebar";
 
 function IdeaLayout() {
   return (
