+++ src/components/Layout/TodoLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { TodoBar } from "../Sidebar/TodoBar";
+import { TodoBar } from "../Sidebar";
 
 function TodoLayout() {
   return (
